contrib: buildable on other platforms (darwin/wins)
Just make it buildable on other platforms.
