+++ contrib/internal/mountns/ns_other.go
@@ -1,0 +1,12 @@
+//go:build !linux
+
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+package mountns
+
+import "fmt"
+
+func Executes(run func() error) error {
+	return fmt.Errorf("not supported")
+}
