+++ contrib/utils/utils_linux.go
@@ -1,0 +1,18 @@
+//go:build linux
+
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+package utils
+
+import (
+	"context"
+	"os/exec"
+	"syscall"
+)
+
+func newExecCommand(ctx context.Context, name string, args ...string) *exec.Cmd {
+	c := exec.CommandContext(ctx, name, args...)
+	c.SysProcAttr = &syscall.SysProcAttr{Pdeathsig: syscall.SIGKILL}
+	return c
+}
