+++ contrib/utils/utils_other.go
@@ -1,0 +1,15 @@
+//go:build !linux
+
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+package utils
+
+import (
+	"context"
+	"os/exec"
+)
+
+func newExecCommand(ctx context.Context, name string, args ...string) *exec.Cmd {
+	return exec.CommandContext(ctx, name, args...)
+}
